Add 'retrieve_cve_dates' stage by hand, as if it was run
This stage precedes the 'retrieve_cve_info' stage that was added with
the help of 'dvc stage add' (and then run with 'dvc repro'), just like
in the https://dvc.org/doc/start/data-management/data-pipelines (it could
be added and run at once with 'dvc run').

To avoid waiting for re-computing what was already computed, the 'dvc.yaml'
and 'dvc.lock' were modified by hand (with the help of 'dvc add' and examining
the contents of *.dvc files).  The 'data/published_cve_df_filename' is no
longer initial data, but an intermediate file, which is tracked by 'dvc.lock',
and not by individual '*.dvc' file.  The 'data/cve_df_filename' was
'dvc add'-ed - it is currently the initial data for the pipeline.  While
at it, update 'data/.gitignore'.

'dvc status' shows that everything is up to date with respect to data pipelines.