Add eval/cve_surv_analysis.py, based in surv_ana.ipynb
It is currently hardcoded to use 'CVSS v3.1 Ratings' column, and with
a hardcoded ranking values (which include also N/A values).

The code was copied from 'notebooks/surv_clean_programming_classes.ipynb',
while the analysis was copied from 'notebooks/surv_ana.ipynb' Jupyter
Notebooks (from the committed version), with some minor changes and
additions.

Note however that code creating survival curve plots is commented out,
and that results of calculation are still gathered into DataFrame
and printed, just like they were displayed in the notebook.

The idea of this script (in the future) is to make it easy to run
different experiments, with different risk factors, and different
parameters... possibly with `dvc exp run`.