cve_surv_analysis.py: Change formatting of printing evaluation measures
Evaluation measures, such as Dxy, are now printed as nested JSON,
instead of storing results in a dataframe (and printing it), as it
was done in the original Jupyter Notebook.  This step is a preparation
on the road to using this script as part of `dvc exp` experiments.

Adds a bit more information to the output.