cve_surv_analysis.py: Add support for --value option
The `--value` option, when given multiple times, can be used to
provide risk column values of interest, and their ranking order.

Useful for columns such as 'access.vector', that have more than 2 unique
non-N/A sring values, but the dataframe does not provide ordering
(in the form of it being categorical ordered column).

.
While at it, fix bug in handling `--lifespan-max` option, namely catching
an exception if this optional parameter is not provided.