cve_surv_analysis.py: Add support for --eval-path option
If this option is provided on the command line, or there is 'eval_path'
key in the 'params.yaml' file, the script would save parameters, metrics
and statistics to files in this directory.

This is preparation for running experiments with DVC (`dvc repro`
or `dvc exp run`, and `dvc params`, `dvc metrics`,...).