dvc: Unset 'path_prefix' for "evaluate-cve_surv" stage, and rerun it
The 'path_prefix' configuration parameter from the params.yaml file
could, if set, make the script write output files in unexpected places
and/or with different filenames than stated in dvc.yaml.  This commit
fixes this issue.

Note that dvc.lock changed more than one would expect because of the
randomness factor involved in the bootstrapping process (different
random seeds; the random seed is not set).