survey_analysis_CVE_metadata.ipynb: Save survival plots
Save all survival function plots in SVG and PDF format with plt.savefig().

Note that for some cases the name of the file must be provided, as
when we do survival analysis for the same risk factor, but with values
dropped or fused.  This feature would be useful for project analysis
in survey_analysis_project_metadata.ipynb.

White at it, re-run the whole notebook.